MemoryDB: refactoring on the memory layout of keys and values (work in progress) and heap allocation strategy
- Reduce the size of the flags to make more room in keys and values
- Use buckets for different page size in the heap
- Added default System Keys in new empty dbs

+++ FoundationDB.Storage.Memory/Core/Memory/EntryPage.cs
@@ -57,7 +57,7 @@ namespace FoundationDB.Storage.Memory.Core
 		public int Count { get { return m_count; } }
 
 		/// <summary>Number of bytes allocated inside this page</summary>
-		public long MemoryUsage { get { return (long)m_current - (long)m_start; } }
+		public ulong MemoryUsage { get { return (ulong)(m_current - m_start); } }
 
 		/// <summary>Type of the entries stored in this page</summary>
 		public abstract EntryType Type { get; }
